INTERLOK-595 - remove the class="" attribute INTERLOK-595 #time 30m
When the SerializableMessage is serialized to XML (or JSON) you get a
either <key-value-pair class="metadata-element"> or "key-value-pair": {
"@class": "metadata-element"...

This is because the underlying translator adds the elements directly
to the KeyValuePairSet.

Made the default translator use the setMetadat(Set) method, and changed
the addMetadata(MetadataElement) method in SerializableMessage to
change the ME into a new KVP before adding it.