INTERLOK-591 #comment initial checkin of interfaces. INTERLOK-591 #time 1h
Initial checkin of interfaces required for a "common" subproject; after
discussions with AM, we have opted to make interfaces that can
be implemented by SerializableMessage. This means we don't have to
move all the stuff from com.adaptris.util and the like in the initial
phase (we may still do that).