INTERLOK-591 WorkflowManagerMBean now extends MessageProcessor INTERLOK-591 #time 2h
WorkflowManager now needs to implement the MessageProcessor interface
@Deprecated existing injectMessage + injectMessageWithReply methods in
WorkflowManagerMBEan. Added tests for those methods in WorkflowManager

Switched to using processAsync in tests that used to use injectMessage.

SerializsedAdaptrisMessage returns a unmodifiable Map when getMessageHeaders()
as we can't easily back the Map<> against the KeyValuePairSet. Probably
need to add in a removeHeader() method into SerializableMessage.