INTERLOK-629 Renamed getters to getTtl() INTERLOK-629 #time 30m
Rather than renaming the member (which really does break config in odd
ways); we renamed the getters and setters to match the member name so
now we have get/setTtl(). Also made it a "Long" which defaults to
null (0).

Deprecated the existing getTimeToLive() method which shouldn't have an
impact on anything other than the Oracle Producer.

renamed ttl member to be timeToLive; because this breaks config
we added in a get/set TTL method as well.