INTERLOK-648 Added in a bootstrap.property validateConfig INTERLOK-648 #time 2h
Added in a bootstrap property validateConfig that defaults to false;
if set to true then it attempts to validate the object post unmarshal.

If the adapter object is invalid; then it throws an exception; because
of where the object is unmarshalled when you start the adapter; it
will actually cause the JVM to effectively system.exit() when starting
up the adapter.