INTERLOK-616 Added a JmxLogger that emits a notification for a logging event. INTERLOK-616 #time 5h
Added logging proxy that can publish a notification to JMX (in the
same way that jmxlogger does). Added an initial log4j12 JmxLogAppender
We should be able to add logback/log4j to implementations when we
decide on what we actually want to convert to.

Uses a blocking queue to queue up the notifications because I haven't
used one for a while, so why not. The loggerId determines the Id that is
used to generate the ObjectName (which is always com.adaptris:Type=Logging)
Example configuration would be :

<appender name="JMX" class="com.adaptris.logging.jmx.log4j12.JmxLogAppender">
  <param name="loggerId" value="MyJmxLogAppender"/>
  <param name="threshold" value="DEBUG"/>
  <layout class="org.apache.log4j.PatternLayout">
    <param name="ConversionPattern" value="%-5p [%t] [%c{1}] %m%n"/>
  </layout>
</appender>

Then we can listen for notifications on
com.adaptris:Type=Logging,id=MyJmxLogAppender