INTERLOK-664 Added a HttpStatusProvider interface INTERLOK-664 #time 5h
Added a HttpStatusProvider interface and implementations which can
derive a dynamic HTTP status.
RawStatusProvider allows you to configure the code as a raw number
ConfiguredStatusProvider allows you to configure it with a typesafe enum
MetadataStatusProvider allows you to configure it from metadata.

Added tests for various providers and modified existing tests to use
the new behaviour.