INTERLOK-668 Ported some components from Apache-HTTP INTERLOK-668 #time 5h
Ported some of the more interesting things out of apache-http.
- ContentTypeProvider to provide a Content-Type field.
- RequestHeaderProvider to provide additional request headers
- ResponseHeaderHandler to handle the HTTP response headers.

Added a StandardHttpProducer which "replaces" JdkHttpProducer and makes
use of the new interfaces. Decided that rather than trying to deprecate
methods and to make the whole thing backwardly compatible, it was a
big enough change to make a whole new class. Some of the things in
JdkHttpProducer were a legacy of the SimpleHttpProducer... Due to the
package changes, apache-http is likely broken.

Marked JdkHttpProducer as deprecated