INTERLOK-668 Added a CompositeResponseHandler INTERLOK-668 #time 1h
Added a CompositeResponseHandler+ ResponseHeaders as ObjectMetadata
along with tests for behaviour.
Changed the behaviour of ResponseHeadersAsMetadata so that the
"null" header field (which appears to be the HTTP status line) is
added with the key "null" (effectively what you get from "" + null).
It's slightly different behaviour to JdkHttpProducer but that's OK.