INTERLOK-674 INTERLOK-682 - DataDestination is now two interfaces. INTERLOK-674 #time 2h INTERLOK-682 #time 4h
The problem is to do with the beautification possibilities of the XML
and subsequently the schema.
If you have a class that contains 2x members both of the same type, then
if you "beautify" the xml (i.e. remove the class="") then you end up
with the possibility of

<xpath-execution>
  <payload-data-destination/>
  <payload-data-destination/>
</xpath-execution>

This causes an invalid schema because we cannot figure out how each of
those classes can be applied to the underlying member fields (source +
target). There is only 1 other class that has this behaviour (Channel)
which was manually worked around by Baldev when he did the original
schema generator.

As a result; separated DataDestination into 2 individual interfaces
DataInputParam + DataOutputParam and have specific implementations of
each interface. DataDestination still exists but extends DIP and DOP.

Made XpathService public once again, and re-enabled the tests.
XPathExecution now requires a DOP<String> and DIP<String> so that the
generic types are enforced properly.