INTERLOK-687 Added a DcoumentBuilderFactoryBuilder INTERLOK-687 #time 1h
Added a builder for configuring a DocumentBuilderFactory instance.
Changed XmlHelper to deprecate most of the other methods, in favour
of the builder parameter. Haven't removed the methods as we need to
understand the impact of this with respect to when we pass a
NamespaceContext around as ObjectMetadata for xpath-alike services.