INTERLOK-677 Log4j2 JMX Appender INTERLOK-677 #time 3h
Added an initial version of a Log4j2 JMX Appender; you can configure
it using :
<JmxLogAppender name="JmxLogAppender">
  <PatternLayout>
    <Pattern>%d{ISO8601} %-5p [%t] [%c{1}.%M()] %m%n</Pattern>
  </PatternLayout>
  <filters>
    <ThresholdFilter level="debug"/>
  </filters>
</JmxLogAppender>

The name is used to build up the object name (just like the loggerId
in the log4j12 impl). The Layout is fixed to be an AbstractStringLayout
rather than Layout<T> so that we can just cast to a string.