INTERLOK-592 Made SASL security only available if protocol=jmxmp INTERLOK-592 #time 1h 30m
Added code that checks the JMXServiceURL for the protocol and ensures
that it is "jmxmp" before setting the sasl callback handler. If you're
trying to use RMI then it's an different ballgame for setting
security. (see the JDMK tutorial). Eventually we may have to have
support that.

Renamed the properties so that they are jmxmp.* rather than sasl.*
Tested with a real adapter, if using jmx+jms (sonicmq); then the jmx
remote profiles is not configured (or the sasl callback). jmxmp contains
the correct environment.