INTERLOK-756 #comment added ability to configure attributes & features INTERLOK-756 #time 4h
Added a transform-factory-attributes and transform-factory-features as
configurable items.
Modified the ErrorListener so that warnings don't include the stack
trace.
<transformer-factory-attributes>
  <key-value-pair>
    <key>http://saxon.sf.net/feature/messageEmitterClass</key>
    <value>net.sf.saxon.serialize.MessageWarner</value>
  </key-value-pair>
</transformer-factory-attributes>
with Saxon-HE logs any xsl:message as warnings in the log file

Added a fail-on-recoverable-error member, which defaults to true (for
backwards compat) which can be turned off if you don't want a recoverable
error to cause an abort.