INTERLOK-865 Changed behaviour of PayloadStreamOutputParam INTERLOK-865 #time 30m
Changed behaviour of PSOP so that if you don't explicitly configure an
encoding; and the message doesn't have an encoding, then we use OutputStream
rather than getWriter(enc). This should "shim" the current broken-ness with
binary payloads.
However the comments on the ticket around deriving the encoding from
the Content-Type header are still pretty sound.