INTERLOK-909 Added MBean operations to retrieve errors. INTERLOK-909 #time 4h
Changed JmxLoggingNotification to capture errors, and preceding lines in a
fairly simplistic FIFO list.
Added MBean methods to expose the error list via JMX and thus to the UI.
Log4j2 configuration allows you to configure the number of lines to keep +
the number of errors to keep. Log4j1.x is a legacy branch so that can stay
at the default.