INTERLOK-983 Added SecurityHAndlerWrapper to EmbeddedConnection INTERLOK-983 #time 1d
If ROOT.war is present (by default it is); then we have to
stop the WAR file, make our changes, and start it. That's because you can't
reconfigure the security handler if the webapp context is already started.
It's all good if you don't have ROOT.war present.
If ROOT.war is present; then we should probably have a single embedded conn
as a shared-component.
Added a test for Embedded+ACL