INTERLOK-1050 Changed interfaces to refer to SerializableMessage not SAM INTERLOK-1050 #time 1h
Changed all the interfaces to use SerializableMessage instead of
SerializableAdaptrisMessage where appropriate. This will break the UI
but we're ok with that.
The only true "break" is if people aren't already using SM as the
return type for those methods; in which case the fix is to do
something like SAM result = new SAM(method(inputParam));