INTERLOK-1071 Added a SafeFileSorter for sorts where File.* could change INTERLOK-1071 #time 2h
Added a SafeFileSorter that only uses file.lastModified() once during
sorting (rather than for every comparison). This should remove any
"IAE - contract broken" exceptions but of course there is the small
chance that we'll end up processing a file out of sequence (even with
quiet-period); but it's no more or less than what it was like during
java6.