INTERLOK-1269 Added code to Workflow to handle cloning INTERLOK-1269 #time 1h
Because most workflows clone the message into a workingCopy we will lose the
object metadata when we get to the message-error-handler. So, put a shim in
to copy the exception information from the working copy into the original.

This is still a bit of a dirty shim, and the whole clone/exception handling
probably needs to be tidied up.