INTERLOK-1345 Changed member variable to be connectionPoolProperties INTERLOK-1345 #time 1h
Changed the member variable to be connectionPoolProperties which should be
emitted correctly by XStream. Can't rename the getter/setter as this will collide
with DatabaseConnection.
This does break existing config, but can't be avoided; this is an edge case as
it doesn't ship with any examples and no-one has commented on how peculiar it
looks as xml...
<connection-properties defined-in="com.adaptris.core.jdbc.DatabaseConnection">