INTERLOK-1455 Added a requestShutdown() method to LoggingConfigurator INTERLOK-1455 #time 2h
Added a requestShutdown() method which (for log4j2) iterates over all the
appenders and asks them to shutdown; LogManager.shutdown() doesn't appear to
ask the appenders to shutdown which means the ExecutorService from the logger
is never terminated.
Log4j12 just does a LogManager.shutdown() - we probably should get rid of this
soon.
Execute the requestShutdown() if we are configcheck only, and also on errors
in the various Bootstrap.main() methods.