INTERLOK-1847 Added a Resolver interface INTERLOK-1847 #time 3h
- Added a Resolver interface and resolver implementations for
system properties & environment variables using
%sysprop{property.name} and %env{env_var_name} respectively.
- In the end, it's probably only useful for very sensitive things
that we never want to be in the adapter.xml (such as passwords); for
things that would change between systest and production you would
still have to use variable substitution, so effectively using %env
is somewhat redundant.

After all:
  my.password=%env{MY_PASSWORD}
is behaviourally the same as
  my.password=${MY_PASSWORD}
But, the 2nd will write the value of password into config; so if
that's not your bag, then %env{} or %sysprop{} exists.