INTERLOK-2350 - Added a convenience password aware metadata filter
Because AdaptrisMessage.getMetadata() returns a shallow clone; when
we modify the metadata, we return a new MetadataElement rather than
modifying the existing MetadataElement.
This will be fine for MetadataFilterService because that explicitly
clears the metadata before re-adding the filtered metadata, but for
producers that embed filters (e.g. jms/http) the decoded password
only exists for the lifetime of the appropriate method which will
be better.