INTERLOK-2916 Enabled non-name FileFiltering.
- Add a 'RemoteFile' that extends File with some of the possible entries
filled in (like length(), lastModified()) so that common "FileFilter"
impls can be used.
- Refactor ApacheFtpClientImpl so that dir(String, FileFilter) applies
the filter on the underlying FTPFile
- Refactor SftpClient so that dir(String, FileFilter) applies the filter
on the underlying SftpChannel.LsEntry
- Enable "overridable enablement" of tests for SftpClient, it re-uses the
"ftp.tests.enabled" property but we want to only test the SftpClient
functionality)
- Existing unit tests aren't broken; so this is a change that hasn't
introduced regression, but could do with explicit testing.
- Removed the legacy FTP client, because commons-ftp FTW (it was only
in stubs; so nothing was using it).