Fix broken test
As of jetty-all.9.4.21.v20190926 the internals of get/setServlet
has changed.
Previously setServlet set _servlet to be what you passed in,
and getServlet would return it ultimately;
Now setServlet doesn't do that exactly, and getServlet returns you
the initialised servlet; which in this case is null
(since we aren't initialising).