INTERLOK-2205 Simplify and optimise
The payload data, even in DefaultAdaptrisMessageImp, can never be null,
as even there the constructor and setPayload() / setContent() methods
set it to an empty byte[] if the users tries to use null. Hence it can
be assumed that whenever the class returns a value to do with the
payload (the payload itself, its length, etc) the payload byte[] won't
be null. This can be enforced (at compile time) with NonNull
annotations.