INTERLOK-2751 Allow XPaths to resolve parts of a payload
Using %payload{xpath:…} allows for parts of an XML payload to be
targeted for extraction. The idea is also to allow JSONPath to do the
same for JSON payloads. It was noticed that this should be in addition
to the original %message{…} resolve and not instead of; likewise with
multi-payload messages and resolving by ID.