INTERLOK-3286 javax validation config checker
- Add a JavaxValidationChecker that uses Validation.buildDefaultValidatorFactory()
- Add a shared hierarchy for DeserializationCheck + JavaxValidationCheck so that
we have common behaviour to get to the Adapter object.
- The config checkers no longer delegate to the AdapterRegistry but instead
manage how to get to the "configuration" themselves. It would have been
better to delegate it to AdapterBuilder but that's a weird rats nest
- Changed the tests so that we don't mock BootstrapProperties since
there is odd behaviour from mockito around returning null after multiple
mock invocations; I suspect that it's because of the class member mocking