INTERLOK-3424 Add emulation of pooling-future-exception-strategy
Add a ServiceErrorHandler interface
ServiceExceptionHandler is the default strategy which means we will throw
an exception (though the aggregator may still throw an exception if you
try to merge XML into JSON)
Add no-exception-if-work-done which behaves like metadata-flag-pfs
Using no-exception-if-work-done implies that you will have a condition
that can "filter" the messages so that you don't try to join a XML doc
to a JSON Array or similar.
Add ignore-all-exceptions for completeness.