INTERLOK-3287 #comment Add a new ConfigDeprecated constraint annotation and its validator class.
The validator will return false if a value is not null (being used).
The ConfigDeprecated annotation should be used with the Deprecated.class as a group:
@ConfigDeprecated(groups = Deprecated.class)

The message will use the template un ValidationMessages.properties but can be overriden.
The removalVersion is used in the default message but can be omitted.