INTERLOK-3465 Don't use a dynamic invalid URL
on Windows the URL contained in config will effectively be
file://localhost/c:/xxx which automagically gets URL encoded by
the helper method. This means, the URL is no longer invalid in
the same way...
Change the URL to be 'hard-coded' to be ./something / with spaces
which will be invalid for the helper.