INTERLOK-3497 - Prep work for verifyReport
Make a new interface out of the "reporting" that happens because of -configcheck
ConsoleReporter is the default service loader implementation (which basically
does what it always did).
This means that verifyReport could just implement ConfigurationReporter and have
a ServiceLoader thing, and then -configcheck would just fire the verify report
to output the Sonar report as required.
Add tests for the default reports.