Add a InputFieldExpressionHelper
Fields are marked as resolvable which means they can be configured such
that we resolve them on a per-message basis. In that situation we can't
preempt performance improvements (like creating a queue on start()
during Producer lifecycle. This helper allows us to determine
if a field is an expression so that components can decide whether to
pre-create things or not.