INTERLOK-3936 Add the concept of payload metadata in multi payload message
They are not really payload metadata as they are available like all other metadata but they are prefixed with PAYLOAD_my-payload-id_ which make them easily accessible in the for each service.
The for each service as also been updated to add the payload id as metadata and to add all payload metadata to the iterated message without the PAYLOAD_my-payload-id_ prefix.
e.g. if the multipayload message has a payload with id my-payload-id and a metadata PAYLOAD_my-payload-id_my-key, when using the foreach the metadata my-key will be also be available making it easier for some use cases.